sync_struct: Add {add,remove}_notifier()
Being able to dynamically (un)publish objects is essential for
some use cases; previously this could only be done by creating
a new Publisher (and thus TCP port) for each dynamically added
notifier, with obvious shortcomings.

This has been battle-tested in production for a considerable
amount of time in the Oxford group as part of an ndscan-related
ARTIQ master extension.