pc_rpc.Server: Marshal exception on unencodable return value
This effectively just moves the pyon.encode() calls into the
try/except block in _process_action().

This way, the client acutally gets an exception if the return
value of a method call is not PYON-serializable instead of the
connection just being dropped, and the exception being swallowed
on the server side.

This is very useful to debug things in practice, as
 1) "… is not PYON-serializable" immediately points physicists who
    just use sipyco as a black box to what the error is, whereas a
    plain ConnectionResetError is rather inscrutable, and
 2) the exception message includes a string representation of the
    intended return value, further making it clear what is going on.