Enable TCP keepalive
Builds on the abandoned existing work.
Makes the settings optional, so that the defaults from the OS will
be used by default. idle=10s, interval=10s and count=3 is quite
aggressive. For our current problem numbers in the minutes will
probably be more appropriate, and it seems inadvisable to bake
these in at this low level.
Adds doc-strings, comments and a sync `create_connection` function.
Also set the keepalive option on accepted connections in
AsyncioServer